+++ lib/common/common/src/bitvec.rs
@@ -6,7 +6,6 @@ use crate::types::PointOffsetType;
 pub type BitOrdering = Lsb0;
 pub type BitVec = bitvec::vec::BitVec<u64, BitOrdering>;
 pub type BitSlice = bitvec::slice::BitSlice<u64, BitOrdering>;
-pub use crate::atomic_bitmask::AtomicBitSlice;
 
 /// Set deleted state in given bitvec.
 ///
