+++ lib/segment/src/index/hnsw_index/hnsw.rs
@@ -5,9 +5,10 @@ use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::thread;
 
 use atomic_refcell::{AtomicRef, AtomicRefCell};
+use common::atomic_bitvec::AtomicBitSlice;
 #[cfg(feature = "gpu")]
 use common::bitvec::BitSlice;
-use common::bitvec::{AtomicBitSlice, BitSliceExt as _, BitVec};
+use common::bitvec::{BitSliceExt as _, BitVec};
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::cow::BoxCow;
 #[cfg(target_os = "linux")]
