+++ lib/segment/src/index/vector_index_search_common.rs
@@ -1,4 +1,4 @@
-use common::bitvec::AtomicBitSlice;
+use common::atomic_bitvec::AtomicBitSlice;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::ScoredPointOffset;
 use itertools::Itertools;
