+++ lib/shard/src/proxy_segment/mod.rs
@@ -5,7 +5,7 @@ pub mod snapshot_entry;
 mod tests;
 
 use ahash::AHashMap;
-use common::atomic_bitmask::AtomicBitVec;
+use common::atomic_bitvec::AtomicBitVec;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
 use itertools::Itertools as _;
